feat: rename ErrorBoundary component
