+++ src/components/common/ErrorBoundary.svelte
